+++ IAttackStrategy.cs
@@ -11,7 +11,6 @@ public class LongRangeAttack : IAttackStrategy
 
     public void Attack(Character character, Enemy enemy)
     {
-        // Check if the attack is a miss
         if (random.NextDouble() < 0.2) // 20% chance to miss
         {
             Console.WriteLine($"{character.Name}'s long-range attack missed!");
@@ -19,7 +18,7 @@ public class LongRangeAttack : IAttackStrategy
         }
 
         // Calculate damage with critical hit chance
-        int damage = character.AttackPower + 5; // Base damage for long range
+        float damage = character.AttackPower + 5; // Base damage for long range
         if (random.NextDouble() < 0.3) // 30% chance for critical hit
         {
             damage *= 2; // Double damage for critical hit
@@ -38,7 +37,7 @@ public class ShortAttack : IAttackStrategy
     public void Attack(Character character, Enemy enemy)
     {
         // Check if the attack is a miss
-        if (random.NextDouble() < 0.05) // 10% chance to miss
+        if (random.NextDouble() < 0.05) // 5% chance to miss
         {
             Console.WriteLine($"{character.Name}'s short-range attack missed!");
             return;
@@ -48,7 +47,7 @@ public class ShortAttack : IAttackStrategy
         float damage = character.AttackPower; // Base damage for short range
         if (random.NextDouble() < 0.2) // 20% chance for critical hit
         {
-            damage *= 1.5; // 50% more damage for critical hit
+            damage *= 1.5f; // 50% more damage for critical hit
             Console.WriteLine("Critical Hit!");
         }
